Removing old files unrelated to concerto. Small refactoring to move viz tools. Cleaning up notebooks and examples.
